cast to int from float divison

+++ sort.py
@@ -23,7 +23,7 @@ def quicksort(arr):
     if len(arr) <= 1:
         return arr
 
-    pivot  = arr[len(arr) / 2]
+    pivot  = arr[int(len(arr) / 2)]
 
     left   = [x for x in arr if x < pivot ]
     middle = [x for x in arr if x == pivot]
